Move auth files to admin/authentication and import into app.modules
 Need to include ngIfs in the html files to not show user admin buttons if not authenticated
