+++ client/src/app/admin/authentication/incorrect-account.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AdminService } from './admin.service';
+import { AdminService } from '../admin.service';
 
 
 @Component({
